@@ -10,40 +10,39 @@
   "dependencies": {
     "@headlessui/react": "^1.5.0",
     "@heroicons/react": "^1.0.6",
-    "@react-aria/utils": "^3.11.3",
-    "@reduxjs/toolkit": "^1.6.2",
-    "axios": "^0.24.0",
-    "axios-auth-refresh": "^3.2.1",
+    "@reduxjs/toolkit": "^1.8.1",
+    "axios": "^0.26.1",
+    "axios-auth-refresh": "^3.2.2",
     "clsx": "^1.1.1",
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
-    "moment": "^2.29.2",
-    "react": "^17.0.0",
+    "moment": "^2.29.3",
+    "react": "^18.0.0",
     "react-activity": "^2.1.3",
-    "react-dom": "^17.0.0",
-    "react-dropzone": "^11.4.2",
-    "react-query": "^3.32.1",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.2",
+    "react-dom": "^18.0.0",
+    "react-dropzone": "^12.0.4",
+    "react-query": "^4.0.0-beta.3",
+    "react-redux": "^7.2.8",
+    "react-router-dom": "^6.3.0",
     "react-use": "^17.3.2",
     "redux": "^4.1.2"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.0",
-    "@types/lodash": "^4.14.176",
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
-    "@vitejs/plugin-react": "1.0.5",
-    "autoprefixer": "^10.4.0",
-    "eslint": "^8.10.0",
-    "eslint-config-prettier": "^8.4.0",
-    "eslint-plugin-react": "^7.29.2",
+    "@types/lodash": "^4.14.181",
+    "@types/react": "^18.0.5",
+    "@types/react-dom": "^18.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.17.0",
+    "@typescript-eslint/parser": "^5.17.0",
+    "@vitejs/plugin-react": "1.3.0",
+    "autoprefixer": "^10.4.4",
+    "eslint": "^8.12.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-react": "^7.29.4",
     "path": "^0.12.7",
-    "postcss": "^8.3.11",
+    "postcss": "^8.4.12",
     "tailwindcss": "^3.0.23",
-    "typescript": "^4.6.2",
+    "typescript": "^4.6.3",
     "vite": "^2.9.1"
   }
 }
