@@ -1,7 +1,6 @@
 import axios from "axios";
 import createAuthRefreshInterceptor from "axios-auth-refresh";
 import React from "react";
-import ReactDOM from "react-dom";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
 import { Provider } from "react-redux";
@@ -9,6 +8,7 @@ import { Api } from "./Api";
 import App from "./App";
 import "./index.css";
 import { store } from "./store";
+import { createRoot } from "react-dom/client";
 
 Api.restoreTokens();
 
@@ -29,12 +29,16 @@ const queryClient = new QueryClient({
   },
 });
 
-ReactDOM.render(
+const container = document.getElementById("root");
+
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!);
+
+root.render(
   <QueryClientProvider client={queryClient}>
     <Provider store={store}>
       <App />
       <ReactQueryDevtools initialIsOpen={false} position="bottom-right" />
     </Provider>
-  </QueryClientProvider>,
-  document.getElementById("root")
+  </QueryClientProvider>
 );
