@@ -1,6 +1,6 @@
-import React, { FC } from "react";
+import React, { FC, ReactNode } from "react";
 
-export const Underlined: FC = ({ children }) => {
+export const Underlined: FC<{ children: ReactNode }> = ({ children }) => {
   return (
     <span className="underline decoration-pink-500 decoration-dotted decoration-2">
       {children}
